agregar funcion de validar role

+++ src/services/UserService.js
@@ -1,18 +1,65 @@
 import { modelUser } from "../models/UserMo.js";
+import { teamModel } from "../models/teamMo.js";
+import { ROLES } from "../constants/roles.js";
 import bcrypt from "bcrypt";
 
 export class UserService {
   constructor() {}
-  async signupUser(data) {
+  invalidRole(newRole, userRoleJWT) {
+    return (
+      newRole == ROLES.SUPER_ADMIN ||
+      (newRole == ROLES.ADMIN && !userRoleJWT == ROLES.SUPER_ADMIN)
+    );
+  }
+  validate(data, userRoleJWT){
+    if (this.invalidRole(data.role, userRoleJWT)) {
+      throw new Error("no tienes permisos para crear este rol");
+    }
+  }
+
+  async createUser({ idTeam, ...data }, userRoleJWT) {
+    this.validate(data, userRoleJWT)
+
+    let team = await teamModel.findById(idTeam);
+    if (!team) {
+      throw new Error("equipo no existe");
+    }
     let user = new modelUser(data);
+
     user.password = await bcrypt.hash(user.password, 10);
-    return await user.save();
+    let newUser = await user.save();
+
+    // delete newUser.password;
+    team.members.push(newUser._id);
+    team.save();
+    return {
+      id: newUser._id,
+      name: newUser.name,
+      email: newUser.email,
+      role: newUser.role,
+    };
   }
-  async deleteUser(id) {
-    await modelUser.findByIdAndDelete(id);
+  async deleteUser(id, userIdJWT) {
+    if (userIdJWT == id) {
+      throw new Error("no puedes eliminar tus propios datos");
+    }
+    return await modelUser.findByIdAndDelete(id);
   }
-  async updateUser(id, data) {
-    await modelUser.findByIdAndUpdate(id, data);
+  async updateUser(id, data, userRoleJWT) {
+    if (data.role) {
+      this.validate(data, userRoleJWT)
+    }
+    let newUser = await modelUser.findByIdAndUpdate(id, data, { new: true });
+    // delete newUser.password;
+    return {
+      id: newUser._id,
+      name: newUser.name,
+      email: newUser.email,
+      role: newUser.role,
+      englishLevel: newUser.englishLevel,
+      knowledge: newUser.knowledge,
+      linkCv: newUser.linkCv,
+    };
   }
   async searchUser(id) {
     let user = await modelUser.findById(id);
@@ -21,5 +68,20 @@ export class UserService {
   async searcAllUsers() {
     let users = modelUser.find();
     return users;
+  }
+  async searchProfile(userIdJWT) {
+    if (!userIdJWT) {
+      throw new Error("el Usuario no esta logeado");
+    }
+    let profile = await modelUser.findById(userIdJWT);
+    return {
+      id: profile._id,
+      name: profile.name,
+      email: profile.email,
+      role: profile.role,
+      englishLevel: profile.englishLevel,
+      knowledge: profile.knowledge,
+      linkCv: profile.linkCv,
+    };
   }
 }
